Separa a implementação do carrossel em um arquivo isolado
